+++ CHANGELOG.md
@@ -32,7 +32,7 @@
 * fix renaming inside `quote` in a `syntax-rules`
 * fix `denominator` of `0`
 * fix `string-ref` wrong exception when index out of range
-* fix `call-with-values` on rejected promise
+* fix ``call-with-values`` with promises
 * fix for Chibi R7RS tests:
   * fix handling an empty symbol `||`
   * fix `char-foldcase` and `boolean=?`